Add optional price table
The table will always exist but whether it's loaded or not
is a configuration setting. If you want to load it, add this
to your config:

    config.add_hook :after_load do |db|
      LedgerWeb::Database.load_prices
    end

Note that this only happens after you change the ledger file.
I still need to work out a good way to trigger a load manually.
Possibly touching the ledger file is good enough?

This commit also adds a "cost" column to the ledger table.